LRSD-8348 [components] Rename Table to ActionTable
