+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/components/ActionTable/index.ts
@@ -1,0 +1,6 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export {default} from './ActionTable';
