+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-custom-element/src/components/index.ts
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
+export {default as Table} from './ActionTable';
 export {default as Badge} from './Badge';
 export {default as Button} from './Button';
 export {default as ButtonDropDown} from './ButtonDropDown';
@@ -13,4 +14,3 @@ export {default as RoundedGroupButtons} from './RoundedGroupButtons';
 export {default as Select} from './Select';
 export {default as Skeleton} from './Skeleton';
 export {default as StatusTag} from './StatusTag';
-export {default as Table} from './Table';
